fix potential uninitialized read
We proactively store `.utf_buf` in case a `PJSON_STATUS_ACCEPT_RETRY` is
returned from the lexer and we have to rollback UTF-8 state. However,
`.utf_buf` was never initiliazed, because the assumption was that
`pj_utf8_push()` will store to it anyway. Turns out that was completely
wrong.

Now always initialize it to `0` in `pjson_init()`.